Move meat of MockHttp to test_openstack.
And rename related fixtures, too.
