+++ test/compute/fixtures/openstack/v1_slug_flavors_detail.xml
