+++ test/compute/fixtures/openstack/v1_slug_images_post.xml
