+++ test/compute/fixtures/openstack/v1_slug_limits.xml
