+++ test/compute/fixtures/openstack/v1_slug_servers_detail_deployment_missing.xml
