+++ test/compute/fixtures/openstack/v1_slug_servers_metadata.xml
