+++ test/compute/fixtures/openstack/v1_slug_shared_ip_groups.xml
