+++ test/compute/test_rackspace.py
@@ -24,6 +24,7 @@ from libcloud.compute.base import Node, NodeImage, NodeSize
 from libcloud.pricing import set_pricing
 
 from test import MockHttp, MockResponse, MockHttpTestCase
+from test.compute.test_openstack import OpenStackMockHttp
 from test.compute import TestCaseMixin
 from test.file_fixtures import ComputeFileFixtures
 
@@ -241,122 +242,11 @@ class RackspaceTests(unittest.TestCase, TestCaseMixin):
         self.assertTrue(self.driver.ex_revert_resize(node=node))
 
 
-class RackspaceMockHttp(MockHttpTestCase):
-
-    fixtures = ComputeFileFixtures('rackspace')
-
-    # fake auth token response
-    def _v1_0(self, method, url, body, headers):
-        headers = {'x-server-management-url': 'https://servers.api.rackspacecloud.com/v1.0/slug',
-                   'x-auth-token': 'FE011C19-CF86-4F87-BE5D-9229145D7A06',
-                   'x-cdn-management-url': 'https://cdn.clouddrive.com/v1/MossoCloudFS_FE011C19-CF86-4F87-BE5D-9229145D7A06',
-                   'x-storage-token': 'FE011C19-CF86-4F87-BE5D-9229145D7A06',
-                   'x-storage-url': 'https://storage4.clouddrive.com/v1/MossoCloudFS_FE011C19-CF86-4F87-BE5D-9229145D7A06'}
-        return (httplib.NO_CONTENT, "", headers, httplib.responses[httplib.NO_CONTENT])
-
-    def _v1_0_UNAUTHORIZED(self, method, url, body, headers):
-        return  (httplib.UNAUTHORIZED, "", {}, httplib.responses[httplib.UNAUTHORIZED])
-
-    def _v1_0_INTERNAL_SERVER_ERROR(self, method, url, body, headers):
-        return (httplib.INTERNAL_SERVER_ERROR, "<h1>500: Internal Server Error</h1>", {}, httplib.responses[httplib.INTERNAL_SERVER_ERROR])
-
-    def _v1_0_UNAUTHORIZED_MISSING_KEY(self, method, url, body, headers):
-        headers = {'x-server-management-url': 'https://servers.api.rackspacecloud.com/v1.0/slug',
-                   'x-auth-token': 'FE011C19-CF86-4F87-BE5D-9229145D7A06',
-                   'x-cdn-management-url': 'https://cdn.clouddrive.com/v1/MossoCloudFS_FE011C19-CF86-4F87-BE5D-9229145D7A06'}
-        return (httplib.NO_CONTENT, "", headers, httplib.responses[httplib.NO_CONTENT])
-
-    def _v1_0_slug_servers_detail_EMPTY(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers_detail_empty.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_servers_detail(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers_detail.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_servers_detail_METADATA(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers_detail_metadata.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
+class RackspaceMockHttp(OpenStackMockHttp):
 
     def _v1_0_slug_flavors_detail(self, method, url, body, headers):
         body = self.fixtures.load('v1_slug_flavors_detail.xml')
         return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_images(self, method, url, body, headers):
-        if method != "POST":
-            raise NotImplemented
-        # this is currently used for creation of new image with
-        # POST request, don't handle GET to avoid possible confusion
-        body = self.fixtures.load('v1_slug_images_post.xml')
-        return (httplib.ACCEPTED, body, {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_images_detail(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_images_detail.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_servers(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers.xml')
-        return (httplib.ACCEPTED, body, {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_EX_SHARED_IP_GROUP(self, method, url, body, headers):
-        # test_create_node_ex_shared_ip_group
-        # Verify that the body contains sharedIpGroupId XML element
-        self.assertTrue(body.find('sharedIpGroupId="12345"') != -1)
-        body = self.fixtures.load('v1_slug_servers.xml')
-        return (httplib.ACCEPTED, body, {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_METADATA(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers_metadata.xml')
-        return (httplib.ACCEPTED, body, {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_72258_action(self, method, url, body, headers):
-        if method != "POST" or body[:8] != "<reboot ":
-            raise NotImplemented
-        # only used by reboot() right now, but we will need to parse body someday !!!!
-        return (httplib.ACCEPTED, "", {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_limits(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_limits.xml')
-        return (httplib.ACCEPTED, body, {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_72258(self, method, url, body, headers):
-        if method != "DELETE":
-            raise NotImplemented
-        # only used by destroy node()
-        return (httplib.ACCEPTED, "", {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_72258_ips(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_servers_ips.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_shared_ip_groups_5467(self, method, url, body, headers):
-        if method != 'DELETE':
-            raise NotImplemented
-        return (httplib.NO_CONTENT, "", {}, httplib.responses[httplib.NO_CONTENT])
-
-    def _v1_0_slug_shared_ip_groups(self, method, url, body, headers):
-
-        fixture = 'v1_slug_shared_ip_group.xml' if method == 'POST' else 'v1_slug_shared_ip_groups.xml'
-        body = self.fixtures.load(fixture)
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_shared_ip_groups_detail(self, method, url, body, headers):
-        body = self.fixtures.load('v1_slug_shared_ip_groups_detail.xml')
-        return (httplib.OK, body, {}, httplib.responses[httplib.OK])
-
-    def _v1_0_slug_servers_3445_ips_public_67_23_21_133(self, method, url, body, headers):
-        return (httplib.ACCEPTED, "", {}, httplib.responses[httplib.ACCEPTED])
-
-    def _v1_0_slug_servers_444222_action(self, method, url, body, headers):
-        if body.find('resize') != -1:
-            # test_ex_resize_server
-            return (httplib.ACCEPTED, "", headers, httplib.responses[httplib.NO_CONTENT])
-        elif body.find('confirmResize') != -1:
-            # test_ex_confirm_resize
-            return (httplib.NO_CONTENT, "", headers, httplib.responses[httplib.NO_CONTENT])
-        elif body.find('revertResize') != -1:
-            # test_ex_revert_resize
-            return (httplib.NO_CONTENT, "", headers, httplib.responses[httplib.NO_CONTENT])
 
 
 if __name__ == '__main__':
